teco-rag/pyproject.toml
[tool.poetry]
authors = ["Vimpas <info@tecorigin.com>"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: BSD License",
  "Operating System :: OS Independent",
]
description = "太初RAG系统"
name = "teco-rag"
readme = "README.md"
version = "1.0.0"
package-mode = false

[[tool.poetry.source]]
name = "tsinghua"
priority = "primary"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"


# 先确保已安装 teco-client-toolkits包
[tool.poetry.dependencies]
python = ">=3.10,<3.12"
nltk = "3.9.1"
pyyaml = "6.0.1"
tqdm = "4.66.2"
numpy = "1.24.4"
pillow = "10.2.0"
torch = "2.4.1"
transformers = "4.45.2"
openai = "1.14.3"
tritonclient = {version = "2.36.0", extras = ["all"]}
sqlalchemy = "2.0.28"
bcembedding = "0.1.3"
fastapi = "0.110.0"
uvicorn = "0.28.0"
pydantic = ">=2.7.4,<3.0.0"
ragas = "0.1.7"
streamlit = "1.32.2"
datasets = "2.18.0"
requests = "2.31.0"
dataclass-wizard = "0.22.3"
python-docx = "^1.1.2"
rapidocr-onnxruntime = "1.3.15"
pymupdf = "1.24.5"
sse-starlette = "1.8.2"
python-multipart = "0.0.9"
unstructured = {extras = ["all-docs"], version = "0.16.4"}
charset-normalizer = "3.2.0"
langfuse = "2.31.0"
wired-table-rec = "0.0.3"
chromadb = "0.5.2"
sentencepiece = "0.2.0"
rank-bm25 = "0.2.2"
pymilvus = {version = "2.4.8", extras = ["model"]}
langchain = "^0.3.4"
teco-client-toolkits = {url = "http://mirrors.tecorigin.com/repository/teco-pypi-repo/packages/teco-client-toolkits/0.0.1/teco_client_toolkits-0.0.1-py3-none-any.whl"}
zhipuai = "^2.1.5.20230904"
langchain-experimental = "^0.3.2"
marker-pdf = "^0.3.10"
jieba = "^0.42.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.3.0"
jupyter = "^1.0.0"

[tool.ruff]
line-length = 96
target-version = "py310"
